feat: icrc2 approve in demo
